refactor: add utils and moved some functions there in fibonacci-page
