+++ src/components/fibonacci-page/fibonacci-page.module.css
@@ -9,7 +9,7 @@
 }
 
 .circles {
-  margin-top: 48px;
+  margin: 48px auto 0;
   display: flex;
   flex-direction: row;
   column-gap: 16px;
